+++ src/client.ts
@@ -1,0 +1,4 @@
+export { ClientDigestAuth } from './lib/client-digest-auth';
+export { ServerDigest } from "./lib/interfaces/client/digest.interface";
+export { GeneratedProtectedResponse, GeneratedUnprotectedResponse } from "./lib/interfaces/client/generated-response.interface";
+export { PayloadProtectionAuth, PayloadProtectionAuthInt, PayloadUnprotected } from "./lib/interfaces/client/payload.interface";
