+++ src/lib/interfaces/client/payload.interface.ts
@@ -3,7 +3,7 @@ export interface PayloadUnprotected {
 }
 
 export interface PayloadProtectionAuth extends PayloadUnprotected {
-  entryBody: string, counter: number, force_algorithm?: string
+  counter: number, force_algorithm?: string
 }
 
 export interface PayloadProtectionAuthInt extends PayloadProtectionAuth {
