+++ src/lib/interfaces/server/generated-response.interface.ts
@@ -1,0 +1,5 @@
+import {ServerDigest} from "./digest.interface";
+
+export interface GeneratedResponse extends ServerDigest {
+  raw: string
+}
