+++ src/lib/interfaces/server/options.interface.ts
@@ -1,0 +1,3 @@
+export interface GenerateResponseOption {
+  domain?: string, opaque?: string, stale?: string, algorithm?: string, qop?:string,
+}
