+++ src/lib/interfaces/server/payload.interface.ts
@@ -1,0 +1,3 @@
+export interface VerifyPayload {
+  entryBody: string, method: string, uri: string
+}
